requirements.txt
# KAN Prosody Prediction Pipeline
# Python 3.9+ required

# Core
numpy>=1.21.0
torch>=2.0.0
pandas>=1.3.0

# Audio/Speech Processing
librosa>=0.9.0
soundfile>=0.10.0
praat-parselmouth>=0.4.0
crepe>=0.0.12

# NLP
spacy>=3.0.0
# Run: python -m spacy download nl_core_news_sm

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Statistics
scipy>=1.7.0

# ASR (optional - for transcription)
openai-whisper>=20230314

# Utilities
tqdm>=4.62.0
openpyxl>=3.0.0  # for reading G2P xlsx files

# MFA (install separately via conda or Docker)
# See: https://montreal-forced-aligner.readthedocs.io/

